Fix colour transformation again, support grid scaling, change user interaction
I accidentally transposed the matrix again.

A hardcoded scaling of `3.464` was used in the triangle grid.
For certain 16x16 textures a different scaling can give better results, where
the stochastic sampling is less aggressive, i.e. patterns in the texture are
better preserved at the cost of slightly visible tiling.

I removed the mouse wheel interaction because it isn't fine for
some reason.